+++ model_navigator/frameworks/torch/utils.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021-2023, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2021-2025, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
 # limitations under the License.
 """Torch utils."""
 
+import gc
 from typing import Optional
 
 from model_navigator.utils.module import lazy_import
@@ -33,3 +34,18 @@ def get_module_device(module: "torch.nn.Module") -> Optional["torch.device"]:
         return next(module.parameters()).device
     except StopIteration:
         return None
+
+
+def offload_torch_model_to_cpu(model: "torch.nn.Module"):
+    """Offload PyTorch model to CPU.
+
+    Args:
+        model: PyTorch model to offload.
+    """
+    model.to("cpu")
+
+    if torch.cuda.is_available():
+        torch.cuda.synchronize()
+        torch.cuda.empty_cache()
+
+    gc.collect()